changes to base api and impl.
git-svn-id: http://mobicents.googlecode.com/svn/trunk@1581 bf0df8d0-2c1f-0410-b170-bd30377b63dc

+++ resources/diameter/base/events/src/main/java/net/java/slee/resource/diameter/base/events/DiameterCommand.java
@@ -37,7 +37,7 @@ public interface DiameterCommand {
      * Return the application ID for this command, e.g., 0
      * @return the application ID for this command
      */
-    int getApplicationId();
+    long getApplicationId();
 
     /**
      * Return the short name for this command, e.g., "CER".
